debug: target size的取整问题

+++ Feature/MultiImages.h
@@ -62,7 +62,7 @@ public:
   vector<vector<Point2f> >   matching_pts;// TODO 前半段为mdlt计算的网格点, 后半段为经过平移计算得到的网格点
 
   /* 图像融合 */
-  Size2f                  pano_size;
+  Size2i                  pano_size;
   vector<Mat>             blend_weight_mask;
 
   /* 接缝线寻找 */
